+++ src/main/java/com/example/algorithm/hash/base/ConsistentHashWithNoVirtualNode.java
@@ -1,4 +1,4 @@
-package com.example.algorithm.hash;
+package com.example.algorithm.hash.base;
 
 import org.springframework.util.StringUtils;
 
@@ -10,77 +10,61 @@ import java.util.UUID;
 
 /**
  * @program: algorithm
- * @description: 带有虚拟节点的一致性hash
+ * @description: 不带虚拟节点的一致性hash
  * @author: pengpeng.wang.o
  * @create: 2019-06-13
  **/
-public class ConsistentHashWithVirtualNode {
+public class ConsistentHashWithNoVirtualNode {
 
     /**
      * 服务器节点信息
      */
     private static SortedMap<Integer,String> nodeMap = new TreeMap<>();
 
-    /**
-     * 虚拟节点数量
-     */
-    private static final int VIRTUAL_NODE_COUNT = 5;
-
     private static String [] servers = {
-            "192.168.56.120:13009",
-            "192.168.56.121:13009",
-            "192.168.56.122:13009",
-            "192.168.56.123:13009",
-            "192.168.56.124:13009"
+        "192.168.56.120:13009",
+        "192.168.56.121:13009",
+        "192.168.56.122:13009",
+        "192.168.56.123:13009",
+        "192.168.56.124:13009"
     };
 
     /**
      * 初始化
      */
-    static{
+    static {
+        System.out.println("hash初始化开始===========");
         for (int i = 0; i < servers.length; i++) {
-            for (int j = 0; j < VIRTUAL_NODE_COUNT; j++) {
-                String server = servers[i];
-                String serverKey = server +"&"+j;
-                nodeMap.put(hash(serverKey),server);
-                System.out.println("hash:"+hash(serverKey));
-            }
+            nodeMap.put(hash(servers[i]),servers[i]);
+            System.out.println("hash取值="+hash(servers[i]));
         }
-        System.out.println("============");
+        System.out.println("hash初始化完成===========");
     }
 
-    /**
-     * hash
-     * @param key
-     * @return
-     */
     public static int hash(String key){
         if (StringUtils.isEmpty(key)){
             return 0;
-        }else {
-            try {
-                MessageDigest md5 = MessageDigest.getInstance("MD5");
-                key = new String(md5.digest(key.getBytes()));
-            } catch (NoSuchAlgorithmException e) {
-                e.printStackTrace();
-            }
-            int hash = 6271;
-            for (int i = 0; i < key.length(); i++) {
-                char c = key.charAt(i);
-                hash+=(hash<<5)+c;
-            }
-            return hash>0?hash:-hash;
         }
+        try {
+            MessageDigest md5 = MessageDigest.getInstance("MD5");
+            key = new String(md5.digest(key.getBytes()));
+        } catch (NoSuchAlgorithmException e) {
+            e.printStackTrace();
+        }
+        int hash = 6667;
+        for (int i = 0; i < key.length(); i++) {
+            char c = key.charAt(i);
+            hash += (hash<<5) + c;
+        }
+        return hash < 0 ? -hash :hash;
     }
-
 
     /**
      * 缓存路由算法
-     * @param key
-     * @return
      */
     public static String getServer(String key){
         int hash = hash(key);
+        System.out.println("hash:"+hash);
         SortedMap<Integer, String> sortedMap = nodeMap.tailMap(hash);
         if (sortedMap.isEmpty()){
             Integer integer = nodeMap.firstKey();
@@ -99,4 +83,5 @@ public class ConsistentHashWithVirtualNode {
             getServer(s);
         }
     }
+
 }
